update(test): print ko case in hub test
