+++ test/integration_test/test_function/print_ng_case.py
@@ -1,0 +1,11 @@
+import glob
+
+
+def print_ng_cases(test_res):
+    if not test_res:
+        return
+
+    for file in glob.glob('ko_case_*_cmd.txt'):
+        with open(file, 'r') as f:
+            print(f.read())
+    return
